feat: include cache creation/read tokens for AWS Bedrock explicit caching (#1721)
**Description**
Include cache creation and cache hit tokens to total input tokens as
well as keep separate fields for cache miss/hit accounting. This is to
unify the usage response to user for both implicit and explicit cache as
the input tokens for gpt and gemini include the cache tokens.

---------

+++ internal/apischema/awsbedrock/awsbedrock.go
@@ -404,11 +404,11 @@ type ConverseOutput struct {
 // TokenUsage is defined in the AWS Bedrock API:
 // https://docs.aws.amazon.com/bedrock/latest/APIReference/API_runtime_TokenUsage.html
 type TokenUsage struct {
-	InputTokens           int  `json:"inputTokens"`
-	OutputTokens          int  `json:"outputTokens"`
-	TotalTokens           int  `json:"totalTokens"`
-	CacheReadInputTokens  *int `json:"cacheReadInputTokens,omitempty"`
-	CacheWriteInputTokens *int `json:"cacheWriteInputTokens,omitempty"`
+	InputTokens           int64  `json:"inputTokens"`
+	OutputTokens          int64  `json:"outputTokens"`
+	TotalTokens           int64  `json:"totalTokens"`
+	CacheReadInputTokens  *int64 `json:"cacheReadInputTokens,omitempty"`
+	CacheWriteInputTokens *int64 `json:"cacheWriteInputTokens,omitempty"`
 }
 
 // ConverseStreamEventType represents a distinct event type received from the Bedrock ConverseStream API.
